The opposite bug screenshot
The filtered angles have two major problems. First, the angle will go in the opposite direction at the beginning of the step response (Figure 4.6.1). This also happened when I was writing the linear Kalman Filter last year. Second, when only one angle is changing, the other one is also changing slightly with it (Figure 4.6.2).